+++ R/dice.R
@@ -130,15 +130,17 @@ dice <- function(data, nk, reps = 10, algorithms = NULL, k.method = NULL,
 #' variables with low signal and (optionally) scales before consensus
 #' clustering. Or, we can use t-SNE dimension reduction to transform the data to
 #' just two variables. This lower-dimensional embedding allows algorithms such
-#' as hierarchical clustering to achieve greater performance.
+#' as hierarchical clustering to achieve greater performance. The largeVis
+#' algorithm is well-suited for visualizing datasets of high dimension by
+#' reducing to a lower-dimensional representation.
 #'
 #' @param data data matrix with rows as samples and columns as variables
 #' @param scale logical; should the data be centered and scaled?
 #' @param type if we use "conventional" measures (default), then the mean and
 #'   standard deviation are used for centering and scaling, respectively. If
 #'   "robust" measures are specified, the median and median absolute deviation
-#'   (MAD) are used. Alternatively, we can apply "tsne" as a method of
-#'   dimension reduction.
+#'   (MAD) are used. Alternatively, we can apply "tsne" or "largevis" as other
+#'   methods of dimension reduction.
 #' @param min.var minimum variability measure threshold used to filter the
 #'   feature space for only highly variable features. Only features with a
 #'   minimum variability measure across all samples greater than \code{min.var}
@@ -155,12 +157,25 @@ dice <- function(data, nk, reps = 10, algorithms = NULL, k.method = NULL,
 #' dim(x)
 #' dim(x.prep)
 prepare_data <- function(data, scale = TRUE,
-                         type = c("conventional", "robust", "tsne"),
+                         type = c("conventional", "robust", "tsne", "largevis"),
                          min.var = 1) {
   type <- match.arg(type)
   if (type == "tsne") {
     return(
-      Rtsne::Rtsne(as.matrix(data), perplexity = 5, max_iter = 500)$Y %>%
+      data %>%
+        as.matrix() %>%
+        Rtsne::Rtsne(perplexity = 5, max_iter = 500) %>%
+        magrittr::extract2("Y") %>%
+        magrittr::set_rownames(rownames(data))
+    )
+  } else if (type == "largevis") {
+    return(
+      data %>%
+        t() %>%
+        largeVis::largeVis() %>%
+        suppressWarnings() %>%
+        magrittr::extract2("coords") %>%
+        t() %>%
         magrittr::set_rownames(rownames(data))
     )
   }